requirements.txt
# Core Dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
PyYAML>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization & Progress
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Optional (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
# Web Server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
websockets>=11.0
streamlit>=1.28.0
pyttsx3>=2.90

